refactor(react-a2a): adopt createRuntimeExtras and split accessor hooks/types
replace the hand-rolled symbolA2AExtras brand + asA2AExtras guard with the shared
createRuntimeExtras helper from @assistant-ui/core/internal; move the three accessor
hooks (useA2ATask/useA2AArtifacts/useA2AAgentCard) into hooks.ts and the
UseA2ARuntimeOptions/UseA2AThreadListAdapter/A2AExtras types into types.ts. the
A2AThreadRuntimeCore controller is untouched and the public barrel is preserved
exactly. the controller rename and the notifyUpdate -> useSyncExternalStore
reactivity rewrite are a deliberate follow-up (behavior-changing, and the current
suite does not cover re-render timing).
