+++ .changeset/a2a-runtime-extras.md
@@ -1,0 +1,5 @@
+---
+"@assistant-ui/react-a2a": patch
+---
+
+refactor: adopt the shared createRuntimeExtras helper and move the accessor hooks and option/extras types into hooks.ts and types.ts, with no public API or behavior change
