+++ packages/react-a2a/src/a2aExtras.ts
@@ -1,0 +1,4 @@
+import { createRuntimeExtras } from "@assistant-ui/core/internal";
+import type { A2AExtras } from "./types";
+
+export const a2aExtras = createRuntimeExtras<A2AExtras>("useA2ARuntime");
