+++ packages/react-a2a/src/useA2ARuntime.ts
@@ -9,116 +9,13 @@ import {
 import type {
   AssistantRuntime,
   AppendMessage,
-  AttachmentAdapter,
-  DictationAdapter,
   ExternalStoreAdapter,
-  ExternalStoreSharedOptions,
-  FeedbackAdapter,
-  RealtimeVoiceAdapter,
-  SpeechSynthesisAdapter,
-  ThreadHistoryAdapter,
   ThreadMessage,
 } from "@assistant-ui/core";
-import { useAuiState } from "@assistant-ui/store";
 import { A2AClient } from "./A2AClient";
-import type { A2AClientOptions } from "./A2AClient";
 import { A2AThreadRuntimeCore } from "./A2AThreadRuntimeCore";
-import type {
-  A2AArtifact,
-  A2AAgentCard,
-  A2ASendMessageConfiguration,
-  A2ATask,
-} from "./types";
-
-// --- Extras symbol for A2A-specific state ---
-
-const symbolA2AExtras = Symbol("a2a-extras");
-
-type A2AExtras = {
-  [symbolA2AExtras]: true;
-  task: A2ATask | undefined;
-  artifacts: readonly A2AArtifact[];
-  agentCard: A2AAgentCard | undefined;
-};
-
-const asA2AExtras = (extras: unknown): A2AExtras => {
-  if (
-    typeof extras !== "object" ||
-    extras == null ||
-    !(symbolA2AExtras in extras)
-  )
-    throw new Error(
-      "This hook can only be used inside a useA2ARuntime provider",
-    );
-  return extras as A2AExtras;
-};
-
-// --- Public hooks for A2A state ---
-
-export const useA2ATask = () => {
-  return useAuiState((s) => asA2AExtras(s.thread.extras).task);
-};
-
-export const useA2AArtifacts = () => {
-  return useAuiState((s) => asA2AExtras(s.thread.extras).artifacts);
-};
-
-export const useA2AAgentCard = () => {
-  return useAuiState((s) => asA2AExtras(s.thread.extras).agentCard);
-};
-
-// --- Thread list adapter type ---
-
-export type UseA2AThreadListAdapter = {
-  threadId?: string;
-  onSwitchToNewThread?: () => Promise<void> | void;
-  onSwitchToThread?: (threadId: string) => Promise<{
-    messages: readonly ThreadMessage[];
-  }>;
-};
-
-// --- Options ---
-
-export type UseA2ARuntimeOptions = ExternalStoreSharedOptions & {
-  /** Pre-built A2A client instance. Provide this OR baseUrl. */
-  client?: A2AClient;
-  /** Base URL of the A2A server. Used to create a client if `client` is not provided. */
-  baseUrl?: string;
-  /** Optional path prefix for all API endpoints (e.g. "/v1"). Does not affect agent card discovery. Only used with baseUrl. */
-  basePath?: string;
-  /** Optional tenant ID for multi-tenant servers. Only used with baseUrl. */
-  tenant?: string;
-  /** Headers for the A2A client (only used with baseUrl). */
-  headers?: A2AClientOptions["headers"];
-  /** A2A extension URIs to negotiate. Only used with baseUrl. */
-  extensions?: string[];
-  /** Extra fetch options (e.g. `{ credentials: 'include' }`). Only used with `baseUrl`. */
-  fetchOptions?: A2AClientOptions["fetchOptions"];
-
-  /** Initial context ID for the conversation. */
-  contextId?: string;
-  /** Default send message configuration. */
-  configuration?: A2ASendMessageConfiguration;
-
-  /** Called when an error occurs. */
-  onError?: (error: Error) => void;
-  /** Called when a run is cancelled. */
-  onCancel?: () => void;
-  /** Called when an artifact is fully received (lastChunk). */
-  onArtifactComplete?: (artifact: import("./types").A2AArtifact) => void;
-
-  adapters?: {
-    attachments?: AttachmentAdapter;
-    speech?: SpeechSynthesisAdapter;
-    dictation?: DictationAdapter;
-    voice?: RealtimeVoiceAdapter;
-    feedback?: FeedbackAdapter;
-    history?: ThreadHistoryAdapter;
-    threadList?: UseA2AThreadListAdapter;
-  };
-};
-
-// --- Main hook ---
+import { a2aExtras } from "./a2aExtras";
+import type { UseA2ARuntimeOptions } from "./types";
 
 export function useA2ARuntime(options: UseA2ARuntimeOptions): AssistantRuntime {
   const [_version, setVersion] = useState(0);
@@ -224,12 +121,11 @@ export function useA2ARuntime(options: UseA2ARuntimeOptions): AssistantRuntime {
       isLoading: core.isLoading,
       messages: core.getMessages(),
       isRunning: core.isRunning(),
-      extras: {
-        [symbolA2AExtras]: true,
+      extras: a2aExtras.provide({
         task: core.getTask(),
         artifacts: core.getArtifacts(),
         agentCard: core.getAgentCard(),
-      } satisfies A2AExtras,
+      }),
       onNew: (message: AppendMessage) => core.append(message),
       onEdit: (message: AppendMessage) => core.edit(message),
       onReload: (parentId: string | null) => core.reload(parentId),
